SGE Update, fullscreen mode, bugfixes
+Added fullscreen mode support (F11)
*Tweaked object scripts to work with 16 bit model inputs and SGE_SetModelAngleInputEvent
*Fixed autorotate checkbox not being checked at the start
*Updated RZUF
*Updated README.md

SGE:
+Added SGE_SetModelAngleInputEvent
*Updated SGE_BMD to version 7:
    *Inputs values are now 16 bit unsigned integers instead of 8 bit
    *Texture sets now have 16 bit texture indexes (up to 65534 textures)
    *All values related to rotation can now be more precise than up to 3deg (full angle value can be specified in INFO block)
    *Every view has its endIndex instead of automatic dividing full angle by views count
*Fixed jagging while rendering offset submodels

+++ src/sge/events/set_model_input.h
@@ -3,15 +3,15 @@
 
 class SGE_SetModelInputEvent : public RZUF3_Event {
 public:
-	SGE_SetModelInputEvent(std::string name, uint8_t value) {
+	SGE_SetModelInputEvent(std::string name, uint16_t value) {
 		m_name = name;
 		m_value = value;
 	}
 
 	std::string getName() const { return m_name; }
-	uint8_t getValue() const { return m_value; }
+	uint16_t getValue() const { return m_value; }
 
 private:
 	std::string m_name;
-	uint8_t m_value;
+	uint16_t m_value;
 };
