+++ src/siudemka/object_scripts/input_by_ui.h
@@ -6,7 +6,7 @@ class RZUF3_UIValueChangeEvent;
 
 class InputByUI : public RZUF3_ObjectScript {
 public:
-	InputByUI(std::string targetObjectName, std::string inputName, uint8_t range = 0);
+	InputByUI(std::string targetObjectName, std::string inputName, uint16_t range = 0);
 	~InputByUI();
 
 	void init();
@@ -14,18 +14,18 @@ public:
 
 	void setTargetObject(std::string targetObjectName);
 	void setInputName(std::string inputName) { m_inputName = inputName; }
-	void setRange(uint8_t range) { m_range = range; }
+	void setRange(uint16_t range) { m_range = range; }
 
 protected:
 	void onUIValueChange(RZUF3_UIValueChangeEvent* event);
 
 	std::string mp_targetObjectName;
 	std::string mp_inputName;
-	uint8_t mp_range;
+	uint16_t mp_range;
 
 	std::string m_targetObjectName = "";
 	std::string m_inputName = "";
-	uint8_t m_range = 0;
+	uint16_t m_range = 0;
 	RZUF3_EventsManager* m_objectEventsManager = nullptr;
 	RZUF3_EventsManager* m_targetEventsManager = nullptr;
