+++ src/siudemka/object_scripts/input_if_ui.h
@@ -6,7 +6,7 @@ class RZUF3_UIValueChangeEvent;
 
 class InputIfUI : public RZUF3_ObjectScript {
 public:
-	InputIfUI(std::string targetObjectName, std::string inputName, int condition, uint8_t valueIfTrue);
+	InputIfUI(std::string targetObjectName, std::string inputName, int condition, uint16_t valueIfTrue);
 	~InputIfUI();
 
 	void init();
@@ -15,7 +15,7 @@ public:
 	void setTargetObject(std::string targetObjectName);
 	void setCondition(int condition) { m_condition = condition; }
 	void setInputName(std::string inputName) { m_inputName = inputName; }
-	void setValueIfTrue(uint8_t valueIfTrue) { m_valueIfTrue = valueIfTrue; }
+	void setValueIfTrue(uint16_t valueIfTrue) { m_valueIfTrue = valueIfTrue; }
 
 protected:
 	void onUIValueChange(RZUF3_UIValueChangeEvent* event);
@@ -28,7 +28,7 @@ protected:
 	std::string m_targetObjectName = "";
 	std::string m_inputName = "";
 	int m_condition = 0;
-	uint8_t m_valueIfTrue = 0;
+	uint16_t m_valueIfTrue = 0;
 	RZUF3_EventsManager* m_objectEventsManager = nullptr;
 	RZUF3_EventsManager* m_targetEventsManager = nullptr;
