+++ src/siudemka/object_scripts/input_if_ui_soft.h
@@ -6,7 +6,7 @@ class RZUF3_UIValueChangeEvent;
 
 class InputIfUISoft : public RZUF3_ObjectScript {
 public:
-	InputIfUISoft(std::string targetObjectName, std::string inputName, int condition, uint8_t targetValue, double step);
+	InputIfUISoft(std::string targetObjectName, std::string inputName, int condition, uint16_t targetValue, double step);
 	~InputIfUISoft();
 
 	void init();
@@ -15,8 +15,8 @@ public:
 	void setTargetObject(std::string targetObjectName);
 	void setCondition(int condition) { m_condition = condition; }
 	void setInputIndex(uint16_t inputIndex) { m_inputName = inputIndex; }
-	void setValueIfTrue(uint8_t targetValue) { m_targetValue = targetValue; }
-	void setStep(uint8_t step) { m_step = step; }
+	void setValueIfTrue(uint16_t targetValue) { m_targetValue = targetValue; }
+	void setStep(uint16_t step) { m_step = step; }
 
 protected:
 	void onUIValueChange(RZUF3_UIValueChangeEvent* event);
@@ -25,13 +25,13 @@ protected:
 	std::string mp_targetObjectName;
 	std::string mp_inputName;
 	int mp_condition;
-	uint8_t mp_targetValue;
+	uint16_t mp_targetValue;
 	double mp_step;
 
 	std::string m_targetObjectName = "";
 	std::string m_inputName = "";
 	int m_condition = 0;
-	uint8_t m_targetValue = 0;
+	uint16_t m_targetValue = 0;
 	double m_step = 0;
 	double m_currentValue = 0;
 	bool m_active = false;
